better thread n determination

+++ include/kreeq.h
@@ -25,7 +25,7 @@ class DBG : public Kmap<UserInputKreeq, DBGkmer, kmer> {
     
     std::queue<std::string*> readBatches;
     std::vector<Buf<kmer>*> buffers;
-    std::vector<uint16_t> buffersDone;
+    std::vector<uint32_t> buffersDone;
     
     std::condition_variable mutexBatches;
